Fixed bug which was not setting default missing value policy for decimal and double signals, added corresponding test: GivenASignal_WhenAddingNewDecimalSignal_NoneQualityMissingValuePolicy_IsDefault, GivenASignal_WhenAddingNewDoubleSignal_NoneQualityMissingValuePolicy_IsDefault, issue #30 resolved

+++ Signals/Domain/Services/Implementation/SignalsDomainService.cs
@@ -38,8 +38,8 @@ namespace Domain.Services.Implementation
             var switchDataType = new Dictionary<DataType, Action>()
             {
                 {DataType.Boolean, ()=>this.missingValuePolicyRepository.Set(signal, new NoneQualityMissingValuePolicy<bool>()) },
-                {DataType.Decimal, ()=>this.missingValuePolicyRepository.Set(signal, new NoneQualityMissingValuePolicy<double>()) },
-                {DataType.Double, ()=>this.missingValuePolicyRepository.Set(signal, new NoneQualityMissingValuePolicy<decimal>()) },
+                {DataType.Decimal, ()=>this.missingValuePolicyRepository.Set(signal, new NoneQualityMissingValuePolicy<decimal>()) },
+                {DataType.Double, ()=>this.missingValuePolicyRepository.Set(signal, new NoneQualityMissingValuePolicy<double>()) },
                 {DataType.Integer, ()=>this.missingValuePolicyRepository.Set(signal, new NoneQualityMissingValuePolicy<int>()) },
                 {DataType.String, ()=>this.missingValuePolicyRepository.Set(signal, new NoneQualityMissingValuePolicy<string>()) }
             };
